bullet now only travel max 200 pixel

+++ 1st SFML Game/entity.h
@@ -17,4 +17,4 @@ public:
 };
 
 float get_angle(Vector2f point1, Vector2i point2); //Get the angle from the entity to a target
-bool point_collide(Vector2f p1, Vector2f p2);
+bool point_collide(Vector2f p1, Vector2i p2);
